+++ src/main/java/subway/presentation/StationController.java
@@ -1,10 +1,14 @@
-package subway;
+package subway.presentation;
 
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.net.URI;
 import java.util.List;
+
+import subway.presentation.request.StationRequest;
+import subway.presentation.response.StationResponse;
+import subway.application.StationService;
 
 @RestController
 public class StationController {
